+++ Engine/StarEntity.h
@@ -8,7 +8,7 @@
 class StarEntity : public Entity
 {
 public:
-	StarEntity(float outerRadius, float innerRadius, int nFlares, const Vec2 _pos = { 0.0f, 0.0f }, Color _c = Colors::Yellow, float pulse = 0.5f)
+	StarEntity(float outerRadius, float innerRadius, size_t nFlares, const Vec2 _pos = { 0.0f, 0.0f }, Color _c = Colors::Yellow, float pulse = 0.5f)
 		:
 		Entity(Star::Make(outerRadius, innerRadius, nFlares), _pos, _c),
 		radius(outerRadius),
